% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.functions.R
\name{importance.of.attributes}
\alias{importance.of.attributes}
\title{importance.of.attributes}
\usage{
importance.of.attributes(ratings, bundles, standardize = 0,
  print.digits = 2)
}
\arguments{
\item{ratings}{a data frame with all clients' ratings}

\item{bundles}{a data frame with all product profiles rated by clients}

\item{standardize}{0, not standardize, 1 o other value, standardize ratings before computing importance}

\item{print.digits}{number of digits to be used when printing results}
}
\value{
conjoint.results       the object returned by this function
}
\description{
This function computes attributes' importance. First I prepare the
list conjoint.results, after that I compute the sample's mean. It uses
treatment effects to estimate part worths; ratings is a matrix of clients (rows) by bundles
rated (colums); bundles is a matrix of bundles (rows) by attributes
(colums). I adapt  code from Marketing Data Science book, chapter 1.
Frist we prepare data
}
\examples{
data(osc)
names(osc)
conjoint.results<-importance.of.attributes(osc$ratings, osc$bundles, 0)
conjoint.results.norm<-importance.of.attributes(osc$ratings, osc$bundles, 1)
}

