% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mktgConjoint.R
\docType{package}
\name{mktgConjoint}
\alias{mktgConjoint}
\alias{mktgConjoint-package}
\title{mktgConjoint: A package for computating the notorious bar statistic.}
\description{
The mktgConjoint package provides three categories of important functions:
mktgConjoint, bar and baz.
}
\section{mktgConjoint functions}{

The mktgConjoint functions ...
}

