% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms.prediction.R
\name{ms.us.conjoint}
\alias{ms.us.conjoint}
\title{ms.us.conjoint}
\usage{
ms.us.conjoint(profiles, ratings, bundles)
}
\arguments{
\item{profiles}{a matrix with the description of competitors' product profiles}

\item{ratings}{a data frame with all clients' ratings}

\item{bundles}{a data frame with all product profiles rated by clients}
}
\value{
ms
}
\description{
This function computes the market share of the product profiles
according to the utility share rule. Uses treatment effects for estimating part worths of profiles;
is the matrix of profiles for which we are going to compute market
profiles is a matrix of product profiles in the market (rows)
by attributes (columns) ratings is a matrix of clients (rows) by
bundles rated (colums) conjointBundles is a matrix of bundles (rows)
by attributes (colums)
}
\examples{
data(osc)
names(osc)
ms.us.conjoint(osc$market.profiles, osc$ratings, osc$bundles)
class(ms.us.conjoint(osc$market.profiles, osc$ratings, osc$bundles))
}

