% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.ms.R
\name{optim.ms.btl}
\alias{optim.ms.btl}
\title{optim.ms.btl}
\usage{
optim.ms.btl(ratings, bundles, exis.profiles, experiment)
}
\arguments{
\item{ratings}{a data frame with all clients' ratings}

\item{bundles}{a data frame with all product profiles rated by clients}

\item{exis.profiles}{a data frame with competitiors' profiles}

\item{experiment}{a data frame with the description of all possible profiles}
}
\value{
optim.list             a list with with the optim profile as well as optim market share
}
\description{
Computes market share for actual profiles as well as for all possible profiles
in order to indentify the bundle that maximazes market share given actual profiles
We use first choice rule
If we combine all three functions into one function, we can use rule=1 for first choice, rule=2 for utility share, and
the rest for BTL. With a if sentence we hace combine all 3 functions
into one.
}
\examples{
library(mktgConjoint)
data(osc)
osc.ms.op.btl<-optim.ms.btl(osc$ratings, osc$bundles, osc$market.profiles, osc$full)
osc.ms.op.btl

}

