% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{osc}
\alias{osc}
\title{Office System Conjoint data set. It is a list with all data sets needed to test the package}
\format{list}
\source{
\url{http://www.decisionpro.biz/}
}
\usage{
osc
}
\description{
Data a conjoint analysis study.
It is fictional data used to illustrate the use of conjoint analysis
to decide the range of products to be offered by an office department store.
The data file osc is a list contening these files:

design:     a data frame with as many rows as attributes used in the conjoint analysis and as many varaibles as levels in each attribute
bundles:    a data frame with as many rows as bundles of profiles indiviuals have rated and as many variables as attributes used in teh analysis
ratings:    a data frame with as many rows as individuals have rated the bundles displayed in the columns
full:       a data frame with a full conjoint design
market.profiles:  a data frame with as many rows as products are in the market (competitors particularly) by attributes (in columns)
constrains: a data frame with some constrains to be used in the estimation of optimal products
reveneu:    a data frame with attributes' variation in cost
}
\examples{
data(osc)
names(osc)
head(names(osc))
}
\keyword{datasets}
\keyword{file}

