% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.functions.R
\name{visualize.importance}
\alias{visualize.importance}
\title{print.attributes.importance}
\usage{
visualize.importance(conjoint.results)
}
\arguments{
\item{conjoint.results}{a  list with the results}
}
\description{
This functions print attributes' importance using R standard plotting system
This functions print attributes' importance using R standard plotting system
}
\examples{
data(osc)
conjoint.results<-importance.of.attributes(osc$ratings, osc$bundles)
visualize.importance(conjoint.results)
}

